+++ Backend/src/main/java/com/example/demo/entities/Event.java
@@ -36,8 +36,8 @@ public final class Event {
     @Column(name = "remaining_players")
     private int remainingPlayers;
     private String location;
-    private Float latitude;
-    private Float longitude;
+    private double latitude;
+    private double longitude;
     private boolean available;
     @ManyToOne
     @JoinColumn(name="user_id", nullable = false)
